requirements.txt
python-telegram-bot==20.7
g4f
deep-translator
python-dotenv
httpx
